worker/requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /tmp/requirements-dev.in
#
attrs==21.2.0
    # via pytest
azure-core==1.14.0
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
azure-eventgrid==4.2.0
    # via -r /tmp/requirements.in
azure-storage-blob==12.8.1
    # via -r /tmp/requirements.in
azure-storage-queue==12.1.6
    # via -r /tmp/requirements.in
boto3==1.17.79
    # via -r /tmp/requirements.in
botocore==1.20.79
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via
    #   minio
    #   msrest
    #   requests
cffi==1.14.5
    # via
    #   cryptography
    #   google-crc32c
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via
    #   azure-storage-blob
    #   azure-storage-queue
    #   msal
    #   pyjwt
flake8==3.9.2
    # via -r /tmp/requirements-dev.in
google-api-core==1.28.0
    # via google-cloud-core
google-auth==1.30.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==1.6.0
    # via google-cloud-storage
google-cloud-storage==1.38.0
    # via -r /tmp/requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.3.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
greenlet==1.1.0
    # via sqlalchemy
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isodate==0.6.0
    # via msrest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
mccabe==0.6.1
    # via flake8
minio==7.0.3
    # via -r /tmp/requirements.in
msal==1.12.0
    # via -r /tmp/requirements.in
msrest==0.6.21
    # via
    #   azure-eventgrid
    #   azure-storage-blob
    #   azure-storage-queue
mypy-extensions==0.4.3
    # via mypy
mypy==0.812
    # via -r /tmp/requirements-dev.in
numpy==1.20.3
    # via pandas
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.9
    # via
    #   google-api-core
    #   pytest
pandas==1.2.4
    # via -r /tmp/requirements.in
pluggy==0.13.1
    # via pytest
protobuf==3.17.1
    # via google-api-core
psycopg2-binary==2.8.6
    # via -r /tmp/requirements.in
py==1.10.0
    # via pytest
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via -r /tmp/requirements.in
pyflakes==2.3.1
    # via flake8
pyjwt[crypto]==2.1.0
    # via msal
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via -r /tmp/requirements-dev.in
python-dateutil==2.8.1
    # via
    #   botocore
    #   pandas
python-json-logger==2.0.1
    # via -r /tmp/requirements.in
pytz==2021.1
    # via
    #   google-api-core
    #   pandas
requests-oauthlib==1.3.0
    # via msrest
requests==2.25.1
    # via
    #   -r /tmp/requirements-dev.in
    #   -r /tmp/requirements.in
    #   azure-core
    #   google-api-core
    #   google-cloud-storage
    #   msal
    #   msrest
    #   requests-oauthlib
rsa==4.7.2
    # via google-auth
s3transfer==0.4.2
    # via boto3
six==1.16.0
    # via
    #   azure-core
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   google-resumable-media
    #   isodate
    #   python-dateutil
sqlalchemy==1.4.15
    # via -r /tmp/requirements.in
toml==0.10.2
    # via pytest
typed-ast==1.4.3
    # via mypy
typing-extensions==3.10.0.0
    # via
    #   mypy
    #   pydantic
urllib3==1.26.4
    # via
    #   botocore
    #   minio
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
